requirements.txt
flask
python-barcode
Pillow
